// plugins/packages/firestore/lib/types.ts
export type SourceOptions = { gcp_key: string };
export type QueryOptions = {
  path: string;
  limit: string;
  where_operation: string;
  where_field: string;
  where_value: string;
  order_field: string;
  order_type: string;
  body: string;
  collection: string;
  records: string;
  operation: string;
};